+++ examples/ct_ecto/lib/ct/tree_path.ex
@@ -13,7 +13,7 @@ defmodule CT.TreePath do
 
   def changeset(path, params \\ %{}) do
     path
-    |> cast(params, [:ancestor, :descendant])
+    |> cast(params, [:ancestor, :descendant, :depth])
     |> validate_number(:ancestor, greater_than_or_equal_to: 0)
     |> validate_number(:descendant, greater_than_or_equal_to: 0)
   end