+++ test/support/author.ex
@@ -1,4 +1,5 @@
 defmodule CTE.Author do
+  @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset
   alias CTE.Comment