+++ test/support/comment.ex
@@ -1,4 +1,5 @@
 defmodule CTE.Comment do
+  @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset