+++ test/support/tree_path.ex
@@ -1,4 +1,5 @@
 defmodule CTE.TreePath do
+  @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset
 
@@ -7,12 +8,12 @@ defmodule CTE.TreePath do
   schema "tree_paths" do
     belongs_to :parent_comment, CTE.Comment, foreign_key: :ancestor
     belongs_to :comment, CTE.Comment, foreign_key: :descendant
-    field :depth, :integer
+    field :depth, :integer, default: 0
   end
 
   def changeset(path, params \\ %{}) do
     path
-    |> cast(params, [:ancestor, :descendant])
+    |> cast(params, [:ancestor, :descendant, :depth])
     |> validate_number(:ancestor, greater_than_or_equal_to: 0)
     |> validate_number(:descendant, greater_than_or_equal_to: 0)
   end